requirements.txt
beautifulsoup4==4.12.3
chromadb==0.4.21
cohere==5.6.2
esprima==4.0.1
langchain==0.2.12
langchain_anthropic==0.1.22
langchain_community==0.2.11
langchain_core==0.2.28
langchain_experimental==0.0.64
langchain_google_genai==1.0.8
langchain_groq==0.1.9
langchain_mistralai==0.1.12
langchain_nvidia_ai_endpoints==0.2.0
langchain_openai==0.1.20
langchain_together==0.1.5
langgraph==0.1.19
libsass==0.23.0
llamaapi==0.1.36
lxml==5.2.2
openai==1.38.0
playwright==1.43.0
psutil==6.0.0
pysqlite3_binary==0.5.3
python-dotenv==1.0.1
retrying==1.3.4
termcolor==2.4.0
thefuzz==0.22.1
todoist_api_python==2.1.5
websocket_client==1.8.0
